Adder functionality basically done
We now mod the result bits by the LCM of the keys, to make them as small
as possible.
The result is no longer truncated.  N bits + N bits = N+1 bit result.
The error levels of the result bits are displayed.
Next going to make this and keygen take command line params, and then
onto the multiplier!